fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,7 +10,7 @@ int main(void)
 {
 	int n;
 
-	long int fib[50];
+	long long int fib[50];
 
 	fib[0] = 0;
 	fib[1] = 1;
@@ -18,7 +18,7 @@ int main(void)
 	for (n = 2; n < 52; n++)
 	{
 		fib[n] = fib[n - 1] + fib[n - 2];
-			printf("%lld, ", fib[n]);
+			printf("%lli, ", fib[n]);
 	}
 	return (0);
 }
